# ewallet-server/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: AstraPay E-Wallet Topup API
  version: v1
  description: Dokumentasi API untuk autentikasi, wallet, transaksi, merchant, dan pengaturan admin.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/users/me:
    get:
      tags: [Pengguna]
      operationId: getProfile
      summary: Melihat profil
      description: Mengambil ringkasan profil customer yang sedang login.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profil berhasil dikembalikan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponseResUserSummaryDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Pengguna]
      operationId: updateProfile
      summary: Memperbarui profil
      description: Memperbarui data profil dasar customer yang sedang login.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReqUpdateProfileDto'
      responses:
        '200':
          description: Profil berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponseResUserSummaryDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/users/me/email:
    post:
      tags: [Pengguna]
      operationId: requestEmailChange
      summary: Meminta perubahan email
      description: Mengirim token verifikasi ke email baru dan mencabut request email sebelumnya.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReqEmailChangeDto'
      responses:
        '200':
          description: Token verifikasi email berhasil dikirim
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponseResEmailChangeDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/users/me/email/confirm:
    post:
      tags: [Pengguna]
      operationId: confirmEmailChange
      summary: Konfirmasi perubahan email
      description: Memvalidasi token email change dan mengganti email customer yang sedang login.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReqEmailChangeConfirmDto'
      responses:
        '200':
          description: Email berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponseResUserSummaryDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/users/me/password:
    put:
      tags: [Pengguna]
      operationId: changePassword
      summary: Mengubah password
      description: Mengganti password customer setelah password lama berhasil diverifikasi.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReqChangePasswordDto'
      responses:
        '200':
          description: Password berhasil diubah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponseVoid'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Payload, query parameter, atau header tidak valid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResponseErrorMap'
    Unauthorized:
      description: Access token atau refresh token tidak valid/tidak ada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResponseVoid'
    Forbidden:
      description: Role pengguna tidak diizinkan mengakses resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResponseVoid'
    NotFound:
      description: Resource tidak ditemukan.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResponseVoid'
    Conflict:
      description: Resource atau request konflik, termasuk idempotency key yang sedang diproses.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResponseVoid'
    InternalServerError:
      description: Kesalahan internal server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResponseVoid'
  schemas:
    BaseResponse:
      type: object
      required: [requestId, status, message]
      properties:
        requestId:
          type: string
          example: req-550e8400-e29b-41d4-a716-446655440000
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          nullable: true
    BaseResponseVoid:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              nullable: true
              example: null
    BaseResponseErrorMap:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              additionalProperties:
                type: string
              example:
                amount: Minimum top-up amount is 10000
    BaseResponseResUserSummaryDto:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ResUserSummaryDto'
...
    BaseResponseResEmailChangeDto:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ResEmailChangeDto'
...
    ReqUpdateProfileDto:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
          example: Budi Santoso
    ReqEmailChangeDto:
      type: object
      required: [newEmail]
      properties:
        newEmail:
          type: string
          format: email
          example: budi.baru@example.com
    ReqEmailChangeConfirmDto:
      type: object
      required: [token]
      properties:
        token:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
    ReqChangePasswordDto:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          type: string
          example: OldPassword1
        newPassword:
          type: string
          minLength: 8
          example: NewPassword1
...
    ResUserSummaryDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Budi Santoso
        email:
          type: string
          format: email
          example: budi@example.com
        createdAt:
          type: string
          format: date-time
          example: '2026-05-18T09:00:00'
        role:
          type: string
          example: CUSTOMER
...
    ResEmailChangeDto:
      type: object
      properties:
        newEmail:
          type: string
          format: email
          example: budi.baru@example.com
        expiresInMinutes:
          type: integer
          format: int32
          example: 15
...